requirements.txt
aiogram==3.20.0
flask
flask-cors
mysql-connector-python

